Integarion fixes
