+++ package.json
@@ -65,6 +65,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "storybook": "7.0.2",
+    "styled-components": "5.3.6",
     "typescript": "4.9.3"
   },
   "peerDependencies": {
@@ -73,8 +74,8 @@
     "styled-components": "^5.3.6"
   },
   "dependencies": {
+    "@taskany/colors": "1.0.4",
     "@monaco-editor/react": "4.4.6",
-    "@taskany/colors": "1.0.3",
     "react-dropzone": "14.2.3",
     "teenyicons": "0.4.1"
   },
