+++ src/components/CardActions.tsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const CardActions = styled.div`
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    box-sizing: border-box;
+    position: absolute;
+    width: 100%;
+    bottom: 0;
+    left: 0;
+    padding: 20px 12px 10px;
+`;
+
+export default CardActions;
