+++ src/components/CardComment.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const CardComment = styled.div`
+    position: relative;
+    box-sizing: border-box;
+    padding: 12px 14px 12px;
+    user-select: auto;
+`;
+
+export default CardComment;
