+++ src/components/CardContent.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const CardContent = styled.div`
+    position: relative;
+    box-sizing: border-box;
+    padding: 12px 14px 60px;
+`;
+
+export default CardContent;
