+++ src/components/index.tsx
@@ -2,6 +2,10 @@ export * from './Badge';
 export * from './Button';
 export * from './Text';
 export * from './Card';
+export * from './CardContent';
+export * from './CardComment';
+export * from './CardActions';
+export * from './CardInfo';
 export * from './CleanButton';
 export * from './ComboBox';
 export * from './Popup';
